Version 0.1.1 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch.
At present, code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.

Bug fixed: Xdata and Ydata were appended to Xdata arraylist only. Now it appends correctly.

Bugs to be fixed: Object mouse does not retain Xdata/Ydata when arraylist is cleared. Have to figure that out.

[Next version 0.2 should contain generation of the measures using the XY position information.]