Version 0.1.2 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 01.08.2019
0.2 version:
Calculates velocity given x y position data as an array.
---------------------------------------------------------------------
Bug fixed: 
1. Xdata and Ydata were appended to Xdata arraylist only. Now it appends correctly.
2. Object mouse does not retain Xdata/Ydata when arraylist is cleared. Now the mouse objects retain their position Xdata and Ydata.
---------------------------------------------------------------------
[Aim: Next version (0.2) should contain generation of the measures using the XY position information.]
---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.