Version 0.2.0 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 02.08.2019
0.2.0 version:
1. Created classes - platform (contains platform properties) and measure (contains measures to be calculated and methods to use on those measures).
2. In class Measures, rDist and RVelocity are calculated correctly. R-Velocity along/perpendicular to platform and R-Velocity Errors are not calculated correctly. Next version should have this rectified.
---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.