Version 0.2.1 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 03.08.2019
0.2.1 version:
1. In class Measures, rDist, RVelocity, RVelocity component along platform, RVelocity component perpendicular to platform, and RVelocity Error Magnitude are calculated correctly.
2. JavaDocs for for Class Measures is incomplete and needs to be updated.
---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.