Version 0.2.2 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 06.08.2019
0.2.2 version:
1. JavaDocs for for Class Measures is completed as of this point.
2. try-catch added to user input on platform details to ensure correct range of numbers are entered.
---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.