Version 0.2.3 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 07.08.2019
0.2.3 version:
1. Maze dimensions added in Class Platform. Allows a value between 0-240.
2. GUI modified into tabs. First tab to select platform and mouse details. Second tab to select measures to be given as output. Only appearance completed. Code to be added.
---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.