Version 0.2.4 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 09.08.2019
0.2.4 version:
1. GUI modified into tabs. First tab to select platform and mouse details. Second tab to select measures to be given as output. Only appearance completed.
HOWEVER, clicking EXTRACT button gives all measures as ASCII files.
IN NEXT VERSION, I will shift all measures into mouse class because measures class complicates creating the output.

---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.