Version 0.3.0 of WM_AnalysisSoftware java code. Created by Meenakshi P. Created from scratch. Updated on 10.08.2019
0.3.0 version:
1. 2nd tab - Check box selection for measures is functional. Radio button ASCII for output format is functional. Code for rest is pending.
2. Class measures has been removed. Have been recast as methods in Class mouse. Makes the code cleaner and easier to write.
3. Errors found in class mouse methods relating to calculating measures. Has been sorted now (problem was to clear the arraylist before adding values at mentioned indices.)


Next task - Generate plots/graphs/images using Java API. [Can fitting be done? Should I use R library for this?]
---------------------------------------------------------------------
Code allows user to select multiple files containing XY position data of a rodent from a Morris Water Maze trial. The user has to select the conditions/settings corresponding to that file(s) i.e. probe day info, genetic background and whether the mouse was injected with a vehicle/saline or a drug to be tested. The files are read and the XY position is stored in a mouse object along with the selected conditions.